pyproject.toml
[build-system]
requires = ["setuptools>=80", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "news-platform"
version = "1.0.0"
description = "Unified news ingestion platform — Telegram channels + Azerbaijani news websites"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  # ── Telegram scraper (news_ingestor) ──
  "alembic>=1.16,<2",
  "asyncpg>=0.30,<1",
  "beautifulsoup4>=4.13,<5",
  "fastapi>=0.115,<1",
  "greenlet>=3.2,<4",
  "httpx>=0.28,<1",
  "lxml>=5.4,<6",
  "orjson>=3.10,<4",
  "pydantic-settings>=2.9,<3",
  "sqlalchemy>=2.0,<3",
  "telethon>=1.42,<2",
  "tenacity>=9,<10",
  "typer>=0.16,<1",
  "uvicorn>=0.34,<1",
  # ── Web scraper (sonxeber_scraper) ──
  "requests>=2.32.0",
  "psycopg>=3.2.0",
  "psycopg-binary>=3.2.0",
  "curl_cffi>=0.10.0",
]

[project.scripts]
news-ingestor = "news_ingestor.cli:app"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src", "."]
testpaths = ["tests"]
